requirements.txt
auto_mix_prep==0.2.0
dm_control==1.0.24
einops==0.8.0
ema_pytorch==0.7.3
extensions==0.4
linear_attention_transformer==0.19.1
matplotlib==3.8.3
mkl==2025.0.0
numpy==1.23.5
opt_einsum==3.3.0
pandas==2.2.3
pykeops==2.2.3
pytorch_lightning==2.2.0.post0
scikit_learn==1.4.1.post1
scipy==1.14.1
seaborn==0.13.2
tensorflow==2.12.0
timm==1.0.11
#torch==2.1.2+cu121
torchvision==0.16.2
tqdm==4.66.5
